+++ package.json
@@ -78,7 +78,8 @@
     "prettier-plugin-organize-imports": "^3.2.2",
     "rimraf": "^5.0.0",
     "typescript": "^5.0.3",
-    "zod": "^3.21.4"
+    "zod": "^3.21.4",
+    "zod-validation-error": "^1.3.0"
   },
   "pnpm": {
     "patchedDependencies": {
