+++ src/config/validateConfig.js
@@ -1,49 +1,65 @@
 import { z } from 'zod'
+import { fromZodError } from 'zod-validation-error'
 
+/** @type {z.ZodType<import('../types.js').GlobConfig>} */
 export const globConfigSchema = z.union([
   z.array(z.string()),
-  z.object({
-    include: z.array(z.string()).optional(),
-    exclude: z.array(z.string()).optional(),
-  }),
+  z
+    .object({
+      include: z.array(z.string()).optional(),
+      exclude: z.array(z.string()).optional(),
+    })
+    .strict(),
 ])
 
-export const cacheConfigSchema = z.union([
-  z.literal('none'),
-  z.object({
+/** @type {z.ZodType<import('../types.js').CacheConfig>} */
+export const cacheConfigSchema = z
+  .object({
     inputs: globConfigSchema.optional(),
     outputs: globConfigSchema.optional(),
     envInputs: z.array(z.string()).optional(),
     usesOutputFromDependencies: z.boolean().optional(),
     inheritsInputFromDependencies: z.boolean().optional(),
-  }),
-])
+  })
+  .strict()
 
-const baseTaskConfigSchema = z.object({
-  runsAfter: z
-    .record(
-      z.object({
-        usesOutput: z.boolean().optional(),
-        inheritsInput: z.boolean().optional(),
-      }),
-    )
-    .optional(),
-  cache: cacheConfigSchema.optional(),
-  parallel: z.boolean().optional(),
-})
+const baseTaskSchema = z
+  .object({
+    runsAfter: z
+      .record(
+        z
+          .object({
+            usesOutput: z.boolean().optional(),
+            inheritsInput: z.boolean().optional(),
+          })
+          .strict(),
+      )
+      .optional(),
+    cache: z.union([z.literal('none'), cacheConfigSchema]).optional(),
+    parallel: z.boolean().optional(),
+  })
+  .strict()
 
-export const topLevelTaskConfigSchema = baseTaskConfigSchema.extend({
-  runType: z.literal('top-level'),
-  baseCommand: z.string(),
-})
+/** @type {z.ZodType<import('../types.js').TopLevelTask>} */
+export const topLevelTaskSchema = baseTaskSchema
+  .extend({
+    runType: z.literal('top-level'),
+    baseCommand: z.string(),
+  })
+  .strict()
 
-export const packageLevelTaskConfigSchema = baseTaskConfigSchema.extend({
-  runType: z.union([z.literal('dependent'), z.literal('independent')]).optional(),
-  baseCommand: z.string().optional(),
-})
+/** @type {z.ZodType<import('../types.js').PackageLevelTask>} */
+export const packageLevelTaskSchema = baseTaskSchema
+  .extend({
+    runType: z.union([z.literal('dependent'), z.literal('independent')]).optional(),
+    baseCommand: z.string().optional(),
+  })
+  .strict()
 
-export const taskConfigSchema = z.union([topLevelTaskConfigSchema, packageLevelTaskConfigSchema])
+/** @type {z.ZodType<import('../types.js').LazyTask>} */
+export const lazyTaskSchema = z.union([topLevelTaskSchema, packageLevelTaskSchema])
 
+/** @type {z.ZodType<import('../types.js').LazyConfig>} */
 export const lazyConfigSchema = z
   .object({
     baseCacheConfig: z
@@ -52,15 +68,28 @@ export const lazyConfigSchema = z
         excludes: z.array(z.string()).optional(),
         envInputs: z.array(z.string()).optional(),
       })
+      .strict()
       .optional(),
-    tasks: z.record(taskConfigSchema).optional(),
+    tasks: z.record(lazyTaskSchema).optional(),
   })
   .strict()
 
 /**
- * @param {Object<*,*>} config
- * @returns {z.infer<typeof lazyConfigSchema>}
+ * @param {import('./resolveConfig.js').LoadedConfig} config
+ * @returns {import('../types.js').LazyConfig}
  */
 export function validateConfig(config) {
-  return lazyConfigSchema.parse(config)
+  try {
+    return lazyConfigSchema.parse(config)
+  } catch (err) {
+    if (err instanceof z.ZodError) {
+      const validationError = fromZodError(err, {
+        issueSeparator: '\n',
+        prefix: '',
+        prefixSeparator: '',
+      })
+      throw new Error(validationError.message)
+    }
+    throw err
+  }
 }
